consenus: Fix raw header encode
