+++ consensus/src/lib.rs
@@ -5,3 +5,19 @@ pub mod argon;
 pub mod dev;
 pub mod difficulty;
 pub mod dummy;
+
+use byteorder::{LittleEndian, WriteBytesExt};
+
+pub fn set_header_nonce(header: &[u8], nonce: u64) -> Vec<u8> {
+    let len = header.len();
+    let mut header = header.to_owned();
+    header.truncate(len - 8);
+    let _ = header.write_u64::<LittleEndian>(nonce);
+    header
+}
+
+pub fn u64_to_vec(u: u64) -> Vec<u8> {
+    let mut wtr = vec![];
+    wtr.write_u64::<LittleEndian>(u).unwrap();
+    wtr
+}
